+++ src/main/java/com/team/comma/domain/alertday/service/AlertDayService.java
@@ -1,7 +1,7 @@
-package com.team.comma.domain.playlist.alertDay.service;
+package com.team.comma.domain.alertday.service;
 
 import com.team.comma.domain.playlist.alertDay.domain.AlertDay;
-import com.team.comma.domain.playlist.alertDay.repository.AlertDayRepository;
+import com.team.comma.domain.alertday.repository.AlertDayRepository;
 import com.team.comma.domain.playlist.playlist.domain.Playlist;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
