+++ src/test/java/com/team/comma/domain/playlist/alertday/AlertDayServiceTest.java
@@ -1,7 +1,8 @@
-package com.team.comma.domain.playlist.alertDay.service;
+package com.team.comma.domain.playlist.alertday;
 
+import com.team.comma.domain.alertday.service.AlertDayService;
 import com.team.comma.domain.playlist.alertDay.domain.AlertDay;
-import com.team.comma.domain.playlist.alertDay.repository.AlertDayRepository;
+import com.team.comma.domain.alertday.repository.AlertDayRepository;
 import com.team.comma.domain.playlist.playlist.domain.Playlist;
 import com.team.comma.domain.playlist.playlist.service.PlaylistService;
 import com.team.comma.domain.user.user.domain.User;
