+++ src/test/java/com/team/comma/domain/playlist/playlist/service/PlaylistServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.doReturn;
 
-import com.team.comma.domain.playlist.alertDay.service.AlertDayService;
+import com.team.comma.domain.alertday.service.AlertDayService;
 import com.team.comma.domain.track.track.service.TrackService;
 import com.team.comma.global.common.dto.MessageResponse;
 import com.team.comma.domain.playlist.playlist.domain.Playlist;
